feat(kyc-requirements): refactor individual requirements
