+++ src/renderer/kyc/application/current-application-popup.jsx
@@ -1,75 +1,15 @@
 import React from 'react';
-import {
-	withStyles,
-	Typography,
-	CircularProgress,
-	FormControlLabel,
-	Checkbox,
-	Grid,
-	Button,
-	Table,
-	TableHead,
-	TableBody,
-	IconButton,
-	FormControl,
-	FormHelperText,
-	Radio,
-	RadioGroup,
-	TableCell
-} from '@material-ui/core';
+import { withStyles, Typography, CircularProgress, Grid, Button } from '@material-ui/core';
 
-import {
-	SmallTableHeadRow,
-	SmallTableCell,
-	MuiEditIcon,
-	MuiAddIcon,
-	SmallTableRow,
-	AttributeAlertIcon,
-	warning,
-	success
-} from 'selfkey-ui';
-
-import { CheckOutlined } from '@material-ui/icons';
+import { KycAgreement } from './kyc-agreement';
+import { KycAgreementText } from './kyc-agreement-text';
+import { KycChecklist } from './kyc-checklist';
 
 import { Popup } from '../../common/popup';
 
 const styles = theme => ({
 	root: {},
 	loading: { textAlign: 'center', paddingTop: '30px' },
-	rowWarning: {
-		color: `${warning} !important;`
-	},
-	warningIcon: {
-		fill: warning
-	},
-	checkIcon: {
-		fill: success
-	},
-	radioGroup: {
-		backgroundColor: 'transparent',
-		marginBottom: 0,
-		paddingTop: '10px'
-	},
-	agreementError: {
-		marginLeft: '30px'
-	},
-	labelColumn: {
-		whiteSpace: 'normal',
-		wordBreak: 'break-all',
-		padding: '5px'
-	},
-	labelCell: {
-		paddingLeft: '5px'
-	},
-	editColumn: {
-		textAlign: 'right'
-	},
-	formControlLabel: {
-		'& span': {
-			fontSize: '14px',
-			lineHeight: '17px'
-		}
-	},
 	link: {
 		color: '#00C0D9',
 		cursor: 'pointer',
@@ -77,220 +17,8 @@ const styles = theme => ({
 	},
 	headCell: {
 		paddingLeft: '15px'
-	},
-	duplicateAddItemBtn: {
-		width: '100px'
-	},
-	duplicateAddItemBtnSmall: {
-		width: '86px',
-		marginTop: '5px',
-		padding: 0
 	}
 });
-
-const KycAgreement = withStyles(styles)(({ text, classes, onChange, value, error }) => {
-	return (
-		<FormControl>
-			<FormControlLabel
-				control={
-					<Checkbox
-						value={value}
-						onChange={(evt, checked) => {
-							onChange && onChange(checked);
-						}}
-					/>
-				}
-				label={text}
-			/>
-			{error && !value ? (
-				<FormHelperText error={true} className={classes.agreementError}>
-					Please confirm you understand what happens with your information
-				</FormHelperText>
-			) : null}
-		</FormControl>
-	);
-});
-
-const KycChecklistItemLabel = withStyles(styles)(
-	({ item, className, classes, selectedAttributes, onSelected, addItem }) => {
-		const { options } = item;
-		if (!options || options.length <= 1) {
-			return (
-				<Typography variant="subtitle1" className={className}>
-					{options.length ? options[0].name : '...'}
-					{item.duplicateType && <br />}
-					{item.duplicateType && (
-						<Button
-							color="primary"
-							size="small"
-							onClick={() => addItem(item)}
-							className={classes.duplicateAddItemBtnSmall}
-						>
-							+ Add Item
-						</Button>
-					)}
-				</Typography>
-			);
-		}
-		const selectedAttr = selectedAttributes[item.uiId] || options[0];
-		onSelected(item.uiId, selectedAttr);
-
-		return (
-			<RadioGroup
-				className={classes.radioGroup}
-				value={selectedAttr.id}
-				onChange={evt =>
-					onSelected(
-						item.uiId,
-						options.find(itm => '' + itm.id === '' + evt.target.value)
-					)
-				}
-			>
-				{options.map(opt => (
-					<FormControlLabel
-						key={opt.id}
-						value={opt.id}
-						control={<Radio />}
-						label={opt.name}
-						className={classes.formControlLabel}
-					/>
-				))}
-				{item.duplicateType && (
-					<Button
-						color="primary"
-						size="small"
-						onClick={() => addItem(item)}
-						className={classes.duplicateAddItemBtnSmall}
-					>
-						+ Add Item
-					</Button>
-				)}
-			</RadioGroup>
-		);
-	}
-);
-
-const KycChecklistItem = withStyles(styles)(
-	({ item, classes, selectedAttributes, onSelected, editItem, addItem }) => {
-		const type = item.title
-			? item.title
-			: item.type && item.type.content
-			? item.type.content.title
-			: item.schemaId;
-		const itemEmpty = !item.options || !item.options.length;
-		const warning = item.required && itemEmpty;
-
-		const warningClassname = warning ? classes.rowWarning : '';
-		const icon = warning ? (
-			<AttributeAlertIcon />
-		) : !itemEmpty ? (
-			<CheckOutlined className={classes.checkIcon} />
-		) : null;
-
-		return (
-			<SmallTableRow>
-				<SmallTableCell className={warningClassname}>{icon}</SmallTableCell>
-				<SmallTableCell>
-					<Typography variant="subtitle1" className={warningClassname}>
-						{type}
-					</Typography>
-				</SmallTableCell>
-				<SmallTableCell className={classes.labelColumn}>
-					<KycChecklistItemLabel
-						item={item}
-						className={warningClassname}
-						selectedAttributes={selectedAttributes}
-						onSelected={onSelected}
-						addItem={addItem}
-					/>
-				</SmallTableCell>
-				<SmallTableCell className={classes.editColumn}>
-					<Typography variant="subtitle1">
-						<IconButton aria-label="Add" onClick={event => addItem(item)}>
-							<MuiAddIcon />
-						</IconButton>
-						{!itemEmpty ? (
-							<IconButton aria-label="Edit" onClick={event => editItem(item)}>
-								<MuiEditIcon />
-							</IconButton>
-						) : null}
-					</Typography>
-				</SmallTableCell>
-			</SmallTableRow>
-		);
-	}
-);
-
-const KycChecklist = withStyles(styles)(
-	({ classes, requirements, selectedAttributes, onSelected, editItem, addItem }) => {
-		return (
-			<Table>
-				<TableHead>
-					<SmallTableHeadRow>
-						<TableCell className={classes.headCell}> </TableCell>
-						<TableCell className={classes.headCell}>
-							<Typography variant="overline">Information</Typography>
-						</TableCell>
-						<TableCell className={classes.labelCell}>
-							<Typography variant="overline">Label</Typography>
-						</TableCell>
-						<TableCell className={classes.editColumn}>
-							<Typography variant="overline">Actions</Typography>
-						</TableCell>
-					</SmallTableHeadRow>
-				</TableHead>
-
-				<TableBody>
-					{requirements.map((item, indx) => {
-						return (
-							<KycChecklistItem
-								item={item}
-								key={indx}
-								selectedAttributes={selectedAttributes}
-								onSelected={onSelected}
-								editItem={editItem}
-								addItem={addItem}
-							/>
-						);
-					})}
-				</TableBody>
-			</Table>
-		);
-	}
-);
-
-const renderPrivacyPolicyText = ({ classes, vendor, purpose, privacyPolicy, termsOfService }) => (
-	<Typography variant="h3">
-		I consent to share my information with {vendor}, for the purposes of {purpose} and that they
-		may further share this information with partners and affiliates in accordance with their{' '}
-		<a
-			className={classes.link}
-			onClick={e => {
-				window.openExternal(e, privacyPolicy);
-			}}
-		>
-			privacy policy
-		</a>{' '}
-		and{' '}
-		<a
-			className={classes.link}
-			onClick={e => {
-				window.openExternal(e, termsOfService);
-			}}
-		>
-			terms and conditions
-		</a>
-		, and by clicking the box, hereby agree to.
-	</Typography>
-);
-
-const renderAgreementText = ({ classes, rpName }) => (
-	<Typography variant="h3">
-		I understand SelfKey Vault will pass this information for <b>Far Horizon Capital Inc</b>,{' '}
-		that will provide {rpName} services in Singapore at my request and will communicate with me{' '}
-		at my submitted email address abore.
-	</Typography>
-);
 
 export const CurrentApplicationPopup = withStyles(styles)(
 	({
@@ -329,19 +57,20 @@ export const CurrentApplicationPopup = withStyles(styles)(
 
 		const purpose = agreement;
 		let description = currentApplication.description;
-		let agreementText = renderPrivacyPolicyText({
-			classes,
-			vendor,
-			purpose,
-			privacyPolicy,
-			termsOfService
-		});
+		let agreementText = (
+			<KycAgreementText
+				vendor={vendor}
+				purpose={purpose}
+				privacyPolicy={privacyPolicy}
+				termsOfService={termsOfService}
+			/>
+		);
 		// Add documents to an existing application
 		if (existingApplicationId) {
 			title = `${rpName || ''} Checklist: ${currentApplication.title}`;
 			description = 'You are required to reupload or provide the missing information:';
 			agreement = true;
-			agreementText = renderAgreementText({ classes, rpName });
+			agreementText = <KycAgreementText vendor="Far Horizon Capital Inc" purpose={rpName} />;
 		}
 
 		const submitDisabled = (agreement && agreementError && !agreementValue) || error;
@@ -403,7 +132,7 @@ export const CurrentApplicationPopup = withStyles(styles)(
 									variant="contained"
 									size="large"
 									onClick={onSubmit}
-									disabled={submitDisabled}
+									disabled={!!submitDisabled}
 								>
 									Submit
 								</Button>
