+++ stories/__fixtures__/kyc-application-data.js
@@ -1,0 +1,16 @@
+export const individualApplication = {
+	rpName: 'relyingPartyName',
+	templateId: 'test-template-id',
+	returnRoute: '/return-route',
+	cancelRoute: '/cancel-route',
+	title: 'Test Process title',
+	description: 'Test process description',
+	agreement: 'test agreement',
+	vendor: 'test-vendor',
+	privacyPolicy: 'test privacy policty',
+	termsOfService: 'test terms of cervice'
+};
+
+export const relyingParty = {
+	name: 'relyingPartyName'
+};
