+++ stories/__fixtures__/loans-data.js
