+++ stories/banking.stories.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { action } from '@storybook/addon-actions';
 import { linkTo } from '@storybook/addon-links';
 import { storiesOf } from '@storybook/react';
-import { resume, country, bankingOffers } from './banking-data';
-import KYCRequirementData from './kyc-requirements-data';
+import { resume, country, bankingOffers } from './__fixtures__/banking-data';
+import KYCRequirementData from './__fixtures__/kyc-requirements-data';
 import BankingOffersTable from '../src/renderer/marketplace/bank-accounts/list/offers-table';
 import BankingAccountTypeTabs from '../src/renderer/marketplace/bank-accounts/list/account-type-tabs';
 import BankingOffersPage from '../src/renderer/marketplace/bank-accounts/list/offers-page';
